Webpage Template
Developed the webpage template for the initial blueprint of all the webpages assorted to the group. This template contains bootstrap, a css file and multimedia (images) file. It is equipped with title section, main header (navbar) section and end credits section (copyright bar). They can utilise this template for header and footer section, and modify the main body accordingly.