Catalogue Page
I have moved the css and js files into their respective folders, in order to make the work more organized.

I have edited the Product Catalogue webpage by adding the 4 products we are going to be selling, which are the neckless page, earrings page, bracelets page and the ring page.