About us page
I have created the about is page. The aim of this page is to give some more information on what the company is and how it all started. ive added several paragraphs of information and implemented a review submission feature. This feature allows users to submit a review and voice their opinions on our website and service.